+++ lib/utils/utils.dart
@@ -4,5 +4,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart' show Color, Colors;
 
+export 'src/api_connector.dart';
+
 part 'src/colors.dart';
 part 'src/layrz_number.dart';